adk-agui/pyproject.toml
[project]
name = "adk-agui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-adk>=0.3.4",
    "fastapi>=0.122.0",
    "google-adk>=1.18.0",
    "litellm>=1.80.5",
    "uvicorn>=0.38.0",
]
